// src/app/shared/enums/shared.enum.ts
export enum MAIN_ROUTES {
  CREATED_LOCATIONS = 'created-locations',
  LOCATION_TYPES = 'location-types',
  ASSIGNED_LOCATION = 'assigned-locations',
  UPLOAD_FILE ='upload-file'
}

export enum LOCATION_TYPE_CATEGORIES {
  EMPLOYEE_LOCATION = 'Employee Location',
  GENERAL_LOCATION = 'General Location'
}

export enum MODE {
  ADD = 'add',
  EDIT = 'edit',
}